lib.rs: Change encryption method case (iOS case sensitivity)
iOS' Camera app seems to be quite picky about the QR string it
detects as WiFi credentials.

Testing on iOS 12.4 reveals that the WiFi connection strings which
have lower cased auth methods (ex: 'wpa2') will fail to be
recognized by the reader.

This change adds a conversion step to the formatter, wherein the
encryption method is always uppercased, before being sent for
QR encoding.

Android's (Lens') detection of WiFi QR codes was not affected
prior to this change and continues to work post-modification.