lib.rs: Add method for printing a QR code directly to the terminal
console_qr() is a very basic method of printing a QR code directly
to the terminal. It does not currently have methods to control
scale, border size, and quiet zone (there isn't one).

For now, it should not be preferred over the other formats provided
in this library (svg, images).